fixed all the issues with prediction and shit goida go
