+++ Content.Goobstation.Client/Security/Systems/ContrabandIconsSystem.cs
@@ -1,4 +1,8 @@
+using Content.Goobstation.Shared.Security.ContrabandIcons.Components;
 using Content.Shared._Goobstation.Security.ContrabandIcons;
+using Content.Shared.Access.Components;
+using Content.Shared.Hands;
+using Content.Shared.Inventory.Events;
 
 namespace Content.Goobstation.Client.Security.Systems;
